+++ CHANGES.rst
@@ -7,6 +7,8 @@ mio 0.1.8.dev
 
 - Implemented ``caller()`` [**functools**]
 - Fixed parser support for ``[1] + [2]`` --> ``[](1} +([](2))``
+- Fixed ``List extend()`` to take another list, not ``*args``.
+
 
 mio 0.1.7 (2013-12-09)
 ......................